#105 - 작품 추천관련
- 작품 요청 타입 enum 전역으로 변경
- 작품추천 전체 화면으로 변경
